+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-10-12
+
+### Fixed
+
+- Updated two world saves that were incorrectly labelled.
+
 ## 2024-10-11
 
 ### Fixed